backend: split rpc_service.py into rpc_client, rpc_registry, and thin shim
rpc_service.py was a 477-line file mixing three unrelated concerns.
Split into focused modules:

- rpc_client.py  — RpcClient class: transport, node queries, block-stats
                   cache, mempool analysis, and performance analytics.
- rpc_registry.py — RpcRegistry, config-file discovery (_find_config),
                   registry construction (_build_registry), and the
                   lazy module-level singleton.
- rpc_service.py — thin public-API shim: re-exports all symbols used by
                   app.py, collector_service, and tests; defines its own
                   _RegistryProxy so patching services.rpc_service._get_registry
                   in tests continues to intercept the registry without
                   also needing to patch rpc_registry.

No behaviour changes; all 71 tests pass.
